aggiunta verifiche e piccoli aggiustamenti

+++ uniscibdtre.sh
@@ -5,7 +5,12 @@ source "./configurazione"
 #-----------------------------------------
 
 cd $uscitaosm
-rm merge.osm
+
+if [ -f merge.osm ]; then
+    rm merge.osm
+else
+    echo "merge.osm NON esiste."
+fi
 
 
 lista=`ls *.osm`
@@ -17,7 +22,7 @@ for i in $lista
       if [ -f "merge.osm" ]
       then
            echo "unisce"
-           $osmosis --rx merge.osm --rx $i --merge --wx temp.osm
+           osmosis --rx merge.osm --rx $i --merge --wx temp.osm
 	   mv temp.osm	merge.osm
       else
            echo "crea merge.osm"
